feat : 커뮤니티 글쓰기 xml, java
